Add submission upload.
The submissions will be uploaded into a submissions array in the hole
document along with their full results.  Currently, the results are not
visible on the site, but that will change soon :)

Had to make a change in the codegolf package to return the full results
so this updates that dependency.