Add user page.
This somewhat undoes the changes from exists => findOne for the user
model that were made a few commits ago.  We definitely do want a
separate action for checking existence/validity of the user than when
fetching a user's details, including submissions.

The main purpose of this commit is to add a page for each user that
lists their scores on each of the holes as well as their total score
across all holes.