Fix bug with user scoreboard.
array_filter does not reset the keys of the array, so element "0" could
have been removed leaving the first element as element "1".  Using
array_values resets the keys to start with 0, and fixes the issue.
Alternatively, rather than accessing element 0, the function "reset"
could be used which returns the first item out of the array.